Move ImdbDataset to data.loader module
